Integration libs expose external event endpoint
What the integration libraries were exposing as a "proactive" endpoint
was not correct. What is now exposed is an external event endpoint that
accepts activities of type `event` and processes them through the bot.

Since an `event` type activity allows a payload to be passed in via
it's `"value"` property, a bot can do whatever it wants. So the bot may
_choose_ to start or continue a conversation proactively, but that is
entirely up any given bot implementation to decide.