Resolve BotFrameworkAdapter from DI if available (WebAPI)
Now attempting to resolve the `BotFrameworkAdapter` as a singleton service from the `HttpConfiguration`'s `DepdencyResolver` during `MapBotFramework<TBot>`. This enables the developer to register the `BotFrameworkAdapter` in their container themselves making it available not only for the integration library to use, but also for direct use in other situations such as injecting it into a custom controller where they want to trigger the bot directly themselves.