Initial commit of reworking of bot AI integration
This is the initial work to directly integration Application Insights
into the Bot Framework such that it doesn't rely on anything external
to the framework such as ASP.NET.

It implements some service extensions which internally register a piece
of Bot Middleware that will start bot a bot framework specific
`TelemetryClient` which will then be enriched with data based on the
`Activity` that is passing through the turn pipeline.

NOTE: as of this commit, several of the unit tests are broken because
they are assuming behaviors from the previous implementation that no
longer exist... and probably won't. I will address those going forward.