Correcting Confirm Prompt and language
ConfirmPrompt has a problem with Language and doesn't recognize the standard for the code of the language. So I put to lower the standard to work with the data from the activity. Like that, the activity send us 'fr-FR' and ConfirmPrompt need to recognize 'fr-fr'. So, I take the 'fr-FR' from the activity and put it to lower to be recognize by the ConfirmPrompt