Added FxCop Analyzers to the AdaptiveExpressions library. (#4278)
* Added FxCop Analyzers to the AdaptiveExpressions library.
Added project level exclusion for CA2225: Operator overloads have named alternates. This should be addressed as a separate issue if we want to support it (see https://docs.microsoft.com/en-us/visualstudio/code-quality/ca2225?view=vs-2019 for additional details).
Added project level exclusion for CA1308: Normalize strings to uppercase. This library assume lowercase across the board.
Added a lot of missing CultureInfo and StringComparision in string conversion calls.
Added comments (the best I could) for all the places where we catch generic exceptions.
Added #pragmas for several things that we can't change because of backward compat.
Added comment and exclude in ExpressionProperty<T> related to CA2214 Do not call overridable methods in constructor (this should not be done but it will require redesigning the class and derived classes).
Replaced calls to LINQ to Count() and Length() by their Count and Length properties where it applied.

* Update AdaptiveExpressions.csproj

Touched file to force build