Addressing suggestions from PR
Removing field to keep track if default client is used. Instead don't set _httpClient (leave null)

This requires additional change to not call AddDefaultRequestHeaders when no custom http client is provided in ctor of BotFrameworkAdapter
Also additional change TO call when we ctor ConnectorClient

Added more detail in comments about why we cannot set retry policy (it does nothing) when we create connector client and we set a custom http client

Changed _httpClient to be called _customHttpClient to help understand that it is custom in usage. Leaving public property HttpClient since that would be breaking change.