Recognizers in Adaptive use DialogContext and Activity (to allow reevaluation of activities after amiguity resolution.)  To adapt to that LuisAdaptiveRecognizer would create a temp turncontext to have the activity in it. The right way to do this is to actually thread the DialogContext/activity through to the luisrecognizer class...
For ExternalEntityRecognizer to work it needs the dc/activity to pass to the external recognizer, so it was time to replace the legacy code with code that threads the dc/activity down to the luis recognizer.

Unit tests for adaptive already go through this path, so no new unit tests are required.