fix: use base64 encoded JSON values
Github actions don't have a lot of rich operators or functions to, for
example, construct JSON.

The `dispatch-workflow` action expects a JSON string that includes some
PR metadata (which may include single or double quotes).

In order to make this as simple as possible, but also as flexible as
possible, base64 encode all JSON values before jamming them in the JSON
string.

The `dispatch-workflow` action, then, knows how to decode these values.