Fixes deliveryMode.ExpectReplies for skills.
When callling /api/messages the return type is always a InvokeResponse<T> data structure.

When a skill receives an activity with deliveryMode=expectReplies, it the botframeworkAdapter caches responses and returns it as a InvokeReponse<ExpectedReplies> meaning
```
{
   "status": status code
   "body": { // type expectedReplies
      "activities": [... ]
    }
}
```
This is correct behavior.

The bug is that the BotFrameworkHttpClient code is interpreting the body of the response as ExpectedReplies instead of InvokeResponse<ExpectedReplies>

The properties don't align and hence not processed.

The unit tests test both sides independently but with mismatched assumptions, so only when e2e is used it doesn't work. (classic metric vs english measurement situation).

This changes client to do correc thing and aligns unit tests to same assumptions as the bot framework adapter class.