Avoid allocations from EvaluationTarget.GetId
GetId is used as part of a loop, and every call to it allocates a new string and possibly invokes Scope.Version(). We can avoid that string allocation by returning the two component pieces as a tuple, and by comparing the individual pieces, if the templates are different we can short-circuit and avoid the Version() invocation.